+++ packages/twopoint5d/CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - _MIGRATION NOTE:_ the `frame` event has been renamed to `renderFrame`
 - the types and constants from `/events.js` are now also included in the main module
   - _MIGRATION NOTE:_ the import of `@spearwolf/twopoint5d/events.js` is now deprecated and should no longer be used. Just use `@spearwolf/twopoint5d` instead.
+- add new constants and types: `OnInitDisplay`, `OnStartDisplay`, `OnRestartDisplay`, `OnPauseDisplay` and `OnDisposeDisplay`
 
 
 ## [0.12.0] - 2025-05-10